# RestService/openapi.yaml
openapi: 3.0.0
info:
  title: iss
  version: 1.0.0
servers:
  - url: http://127.0.0.1:5005
  - url: http://localhost:5005
paths:
  /api/datatransfer:
    delete:
      tags:
        - default
      summary: delete
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: recordid
          in: query
          schema:
            type: integer
          example: '0'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    put:
      tags:
        - default
      summary: put
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                _id: 0
                Bytes: 1
                Packets: 55
                Timestamp: 1713771082
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: recordid
          in: query
          schema:
            type: integer
          example: '0'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    post:
      tags:
        - default
      summary: post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                _id: 0
                Bytes: 1
                Packets: 55
                Timestamp: 1713771082
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: recordid
          in: query
          schema:
            type: integer
          example: '0'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    get:
      tags:
        - default
      summary: get
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: recordid
          in: query
          schema:
            type: integer
          example: '1'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/datatransfermax:
    get:
      tags:
        - default
      summary: min
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: timestamp
          in: query
          schema:
            type: integer
          example: '0'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/datatransferavg:
    get:
      tags:
        - default
      summary: avg
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: timestamp
          in: query
          schema:
            type: integer
          example: '0'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/datatransfersum:
    get:
      tags:
        - default
      summary: sum
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: timestamp
          in: query
          schema:
            type: integer
          example: '0'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}